[fx] Flatten all lists in matcher
To match nodes within the graph, the matcher currently flattens the
arguments and compares each argument against each other. However, if it
believes that a list input contains all literals, it will not flatten
the list and will instead compare the list directly against each other.
It determines if a list is a literal by checking if the first element is
a node.

However, there are cases where we want to check all the values in the
list (such as in the test case), so I think it might just be simpler to
flatten all lists.